refactor(tui): split app/mod.rs into state, selection, and jump modules
app/mod.rs had grown to 1056 lines (over ADR 0028's 1000-line warn
threshold), mixing three independent responsibilities: App's own field
layout/construction/plain accessors, the selected_* cursor-row
resolution queries, and the jump-popup/jumplist feature. Split each
into its own sibling module (state.rs, selection.rs, jump.rs),
mirroring the existing handle_key.rs precedent (PR #82's
app/{mod,input_key,tests} split) of a dispatch-only file plus
struct/accessors staying in mod.rs.

Pure code motion: every moved function body is unchanged. App's
public API is unchanged (all pub items re-exported from app/mod.rs).
Struct fields and a few jumplist internals widen from module-private
to pub(super) so handle_key.rs (a mod.rs sibling) can keep reaching
them across the new module boundary.
